+++ quotes/stock_security_list.go
@@ -7,7 +7,7 @@ import (
 )
 
 const (
-	TDX_SECURITY_LIST_MAX = uint16(1000) // 单次最大获取多少条股票数据
+	TDX_SECURITY_LIST_MAX = 1000 // 单次最大获取多少条股票数据
 )
 
 // SecurityListPackage 股票列表
